+++ internal/store/transaction_store.go
@@ -136,8 +136,8 @@ func (m *transactionModel) GetAll(userId int64, title string, tags []string, sta
 	query := fmt.Sprintf(`SELECT id, user_id, account_id, type, title, description, tags, amount, payday, created_at, version
 	FROM transactions
 	WHERE user_id = $1 
-	AND ((to_tsvector('simple', title) @@ to_tsquery('simple', $2) OR $2='')
-	OR (tags @> $7 OR $7 = '{}'))
+	AND (to_tsvector('simple', title) @@ to_tsquery('simple', $2) OR $2='')
+	AND (tags @> $7 OR $7 = '{}')
 	AND payday >= $3 AND payday < $4
 	ORDER BY %s %s, id ASC
 	LIMIT $5 OFFSET $6`, filters.sortColumn(), filters.sortDirection())
